Move shader.dart into shaders directory
