+++ lib/web_ui/lib/src/engine.dart
@@ -88,7 +88,7 @@ part 'engine/html/recording_canvas.dart';
 part 'engine/html/render_vertices.dart';
 part 'engine/html/scene.dart';
 part 'engine/html/scene_builder.dart';
-part 'engine/html/shader.dart';
+part 'engine/html/shaders/shader.dart';
 part 'engine/html/surface.dart';
 part 'engine/html/surface_stats.dart';
 part 'engine/html/transform.dart';
